RAG_Production_APP/backend/requirements.txt
fastapi
inngest
llama-index-core
llama-index-readers-file
python-dotenv
qdrant-client
uvicorn
streamlit
google-generativeai
langchain-google-genai
slack_sdk
redis
